Refactor RTSP server to use hosted service pattern
- Updated `Program.cs` to implement `Host` and `IHostedService`, moving RTSP server initialization to `RTSPServerWorker`.
- Changed project file to use `Microsoft.NET.Sdk.Worker` and updated package references.
- Created `RTSPServerWorker.cs` to manage the RTSP server lifecycle, including configuration and logging.
- Modified `appsettings.json` to nest RTSP server settings under `RTSPServerApp` and added a `Logging` section.
