+++ src/RTSPServerApp/appsettings.json
@@ -1,7 +1,14 @@
 ﻿{
-  "HostName": "127.0.0.1",
-  "Port": 8554,
-  "FilePath": "frag_bunny.mp4",
-  "UserName": "admin",
-  "Password": "password"
+  "RTSPServerApp": {
+    "HostName": "127.0.0.1",
+    "Port": 8554,
+    "FilePath": "frag_bunny.mp4",
+    "UserName": "admin",
+    "Password": "password"
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Debug"
+    }
+  }
 }
